refactor: Move `MeasureUnit` to `measure` crate
