+++ components/experimental/src/measure/mod.rs
@@ -4,3 +4,5 @@
 
 pub(crate) mod power;
 pub(crate) mod si_prefix;
+
+pub mod measureunit;
