+++ provider/source/src/units/info.rs
@@ -6,10 +6,8 @@ use std::collections::{BTreeMap, HashSet};
 
 use crate::SourceDataProvider;
 use crate::{cldr_serde, units::helpers::ScientificNumber};
-use icu::experimental::units::{
-    measureunit::MeasureUnitParser,
-    provider::{ConversionInfo, UnitsInfoV1, UnitsInfoV1Marker},
-};
+use icu::experimental::units::provider::{ConversionInfo, UnitsInfoV1, UnitsInfoV1Marker};
+use icu_experimental::measure::measureunit::MeasureUnitParser;
 use icu_provider::prelude::*;
 use zerotrie::ZeroTrieSimpleAscii;
 use zerovec::VarZeroVec;
